% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi.R
\name{estimate_pi}
\alias{estimate_pi}
\title{Simple Pi estimation}
\usage{
estimate_pi(B = 5000, seed = 10)
}
\arguments{
\item{B}{A \code{numeric} (integer) used to denote the number of points simulated}

\item{seed}{A \code{numeric} used to control the seed of the random number
generator used by this function.}
}
\value{
A \code{list} of class pi with the attributes points and estimation
}
\description{
returns  an estimated value of π and simulated points
}
\examples{
estimate_pi(500, 6)
}
