% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi.R
\name{plot.pi}
\alias{plot.pi}
\title{Plot of pi}
\usage{
plot.pi(x)
}
\arguments{
\item{x}{A \code{lits} of class pi returned by the function simulate_pi}
}
\value{
A \code{list} of class pi with the attributes
}
\description{
Plot of pi
}
\section{Functions}{
\itemize{
\item \code{plot.pi}: simulated points
}}

